Title: 자연수 n개의 합, Time: 111ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 250604/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -7,7 +7,7 @@ public class Main {
     public static void main(String[] args) throws IOException {
         Scanner sc = new Scanner(System.in);
 
-        long s = sc.nextInt();
+        long s = sc.nextLong();
 
         long left = 1;
         long right = (long) Math.pow(2, 18);
